# apps/server/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lina account API",
    "version": "0.1.0",
    "description": "Bun + TypeScript account service. All mutating API calls require an exact trusted Origin header, including CLI clients. Session cookies are HTTP-only. This contract contains no server credentials. Email verification links and reset callbacks are handled by Better Auth 1.7.4; use links from email as opaque values."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3002",
      "description": "Local development; configure HTTPS in production"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-better-auth.session_token"
      },
      "developmentSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    }
  },
  "paths": {
    "/api/v1/me": {
      "get": {
        "summary": "Read your identity and access",
        "tags": [
          "Accounts"
        ],
        "parameters": [],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "developmentSessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "name": {
                          "type": "string"
                        },
                        "emailVerified": {
                          "type": "boolean"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "member",
                            "admin",
                            "owner"
                          ]
                        },
                        "twoFactorEnabled": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "name",
                        "emailVerified",
                        "role",
                        "twoFactorEnabled"
                      ]
                    },
                    "access": {
                      "type": "object",
                      "properties": {
                        "allowed": {
                          "type": "boolean"
                        },
                        "reason": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "tier": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "full_access",
                                "orchestrator"
                              ]
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "features": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "expiresAt": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date-time"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "revision": {
                          "type": "integer"
                        },
                        "accountStatus": {
                          "type": "string"
                        },
                        "serverTime": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "allowed",
                        "reason",
                        "tier",
                        "features",
                        "expiresAt",
                        "revision",
                        "accountStatus",
                        "serverTime"
                      ]
                    }
                  },
                  "required": [
                    "user",
                    "access"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "403": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "413": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "429": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "503": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/sessions/{id}": {
      "delete": {
        "summary": "Revoke an owned session with recent authentication",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "developmentSessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "403": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "413": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "429": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "503": {
            "description": "See stable error code; authentication library errors retain their own 4xx codes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  }
}